Property Price Prediction Test and Train data.
The two files contain the raw data used for applying linear regression to predict the Sale_Price of a house.
The train data is where we will train the model using R software.
Using the optimum linear model developed by training, we will then predict the house prices for the Test dataset.